Fix init to run on actual RCX instead of simulator
So it turns out I didn't take across a critical block of code labeled
__init from the old wrapper, which appears to be emptying BSS and
overwriting the firmware. The linker script I'd been using from BrickOS
doesn't have the sections that match this __init ASM block, so I've gone
back to the original linker script too. We can keep the nicer way of
objcopying the coff to the srec instead of building twice though...
The emulator (brickemu) didn't care about the missing init, most likely
because the registers and firmware are cleared automatically... The
perils of testing on an emulator :)