Refactoring, interfacing to librcx in tvm-rcx
The idea of this structure is that code from librcx will be brought in
verbatim to librcx.c, and the appropriate C abstractions over the low
level functions will reside in tvm-rcx.c - it may be later that
functions to be FFI wrapped will need to reside separately from tvm-rcx,
but for now they will live there too. This should keep librcx/RCX
low-level code out of tvm.c, to keep the wrapper as clean as possible.