Link against the TVM succesfully, change host
This brings in headers and calls the context setup function in the TVM,
causing libtvm to be linked in. This builds as a 37K srec, which given
that the final firmware previously was 39K may be a problem. Also
changed to using the h8300-hms host setting rather than
h8300-hitachi-coff. I don't think it matters in retrospect, but hms *is*
more correct. Also, turn off the memory allocator for now.